updated plotting
Added ability to plot ellipses and hulls around each group of data, in
addition to hulls between the centres of groups of community members.

Implemented Andrew Parnell’s code for transforming a unit circle to an
ellipse based on the covariance matrix. This also allows prediction
ellipses containing x% of data to be specified in addition to the
standard ellipse which contains ~40% of the data.